Completely rework the meaning of LValues in the compiler
They now have nothing to do with mutability - they just mean "this represents an object in memory".
I had to rework the way the typechecker decides how it handles
mutability for this. This is a pretty major rework, and I'll try to
explain how the new rules work in the docs later.