DeclRefs of FuncDecls no longer have an LValue type.
This decision is intended to make it easier to generate the IR. Having an LValue for those do not make sense, as they're sort of "compile time constant pointers to a function's code", not really an object that we can load. LValues are loadable by nature, and thus do not fit FuncDecl.